Make Rust NumPy batches zero-copy
Borrow slab-backed batch memory into NumPy arrays and cast bucket indices to int64 in Rust so Python can consume ready batches without an extra host-side copy.

+++ loader/python/nnue_loader/_internal.pyi
@@ -15,8 +15,8 @@ class Batch(TypedDict):
     score: npt.NDArray[np.float32]
     white: npt.NDArray[np.int32]
     black: npt.NDArray[np.int32]
-    psqt_indices: npt.NDArray[np.int32]
-    layer_stack_indices: npt.NDArray[np.int32]
+    psqt_indices: npt.NDArray[np.int64]
+    layer_stack_indices: npt.NDArray[np.int64]
 
 class PipelineStats(TypedDict):
     decoded_entries: int
